Enforce clientId timeout in clientExists
If there is a timeout set check that an existing clientId's timestamp is
newer than double that timeout. If no timeout is set just check for the
clientId's existence.

This mitigates a race condition that is triggered when a connect and/or
subscribe occurs for a given clientId that is in the midst of being reaped
by the garbage collector. In this case orphaned messages queues and channel
subscriptions could remain after a clientId was reaped.