Don't destroy clients when publishing
Publishing was stepping on GC's toes by calling `destroyClient` whenever
it detected a client ID was past its prime. I *think* there may have
been some contention between the two (publisher and gc), so this moves
the cleanup responsibility wholly to the latter.

Well, not completely. The `disconnect` function inside of Faye itself
calls `destroyClient` as well, but let's leave that be.