Rewrite destroyClient to tolerate Redis failures
A lot of changes here. Primarily, the destroyClient function now
tolerates Redis failures all the way downstream, and if any error is
detected, the destroy process stops. Instead of trying to recover, we
rely on the hope that GC will pick up a client and re-run the calls in
the future, since the removal from the /clients sorted set is the final
Redis call.

The GC lock has also been removed. After grabbing a list of expired
client IDs, we iterate over them with destroyClient, which returns
immediately. Since we're simply launching destroyClient into flight, and
don't care if it succeeds or not, we allow GC to re-run in the future
without any restrictions.

Because destroyClient doesn't block, we have an additional, expected
test failure -- "Redis engine destroyClient when the client has
subscriptions stops the client receiving messages:". This will fail b/c
it's written to expect the destroy to occur immediately. If the test is
rewritten with the expectation in a callback, it succeeds.