Set the score with the initial ZADD in createClient
Previously, it would set the initial score to zero, and then allow
ping() to update the score to now in the callback. This seems like it
would allow a window of opportunity for GC to sneak in and kill the
newly created client, however, so now it simply sets the score in the
initial call.